+++ moco-core/src/main/java/com/github/dreamhead/moco/matcher/PathMatcher.java
@@ -1,7 +1,10 @@
 package com.github.dreamhead.moco.matcher;
 
+import com.github.dreamhead.moco.Request;
 import com.github.dreamhead.moco.RequestExtractor;
 import com.github.dreamhead.moco.RequestMatcher;
+import com.github.dreamhead.moco.internal.ContextKey;
+import com.github.dreamhead.moco.internal.SessionContext;
 import com.github.dreamhead.moco.resource.Resource;
 import com.github.dreamhead.moco.util.AntPathMatcher;
 
@@ -10,9 +13,14 @@ public class PathMatcher<T> extends AbstractOperatorMatcher<T> {
 
     public PathMatcher(final RequestExtractor<T> extractor, final Resource expected) {
         super(extractor, expected, input -> {
-            final String target = expected.readFor(null).toString();
+            final String target = expected.readFor((Request) null).toString();
             return matcher.match(target, input);
         });
+    }
+
+    @Override
+    public void attach(final SessionContext context) {
+        context.register(ContextKey.PATH, getExpected());
     }
 
     @Override