+++ moco-core/src/test/java/com/github/dreamhead/moco/MocoTemplateTest.java
@@ -6,11 +6,11 @@ import com.google.common.collect.Iterables;
 import com.google.common.io.ByteStreams;
 import com.google.common.io.Resources;
 import org.apache.hc.client5.http.HttpResponseException;
+import org.apache.hc.client5.http.fluent.Request;
 import org.apache.hc.core5.http.Header;
 import org.apache.hc.core5.http.HttpResponse;
 import org.apache.hc.core5.http.HttpVersion;
 import org.apache.hc.core5.http.ProtocolVersion;
-import org.apache.hc.client5.http.fluent.Request;
 import org.apache.hc.core5.http.message.BasicNameValuePair;
 import org.junit.Test;
 
@@ -19,8 +19,8 @@ import java.net.URL;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.text.SimpleDateFormat;
-import java.util.Date;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
 
 import static com.github.dreamhead.moco.Moco.and;
 import static com.github.dreamhead.moco.Moco.by;
@@ -29,6 +29,7 @@ import static com.github.dreamhead.moco.Moco.eq;
 import static com.github.dreamhead.moco.Moco.file;
 import static com.github.dreamhead.moco.Moco.header;
 import static com.github.dreamhead.moco.Moco.jsonPath;
+import static com.github.dreamhead.moco.Moco.path;
 import static com.github.dreamhead.moco.Moco.pathResource;
 import static com.github.dreamhead.moco.Moco.status;
 import static com.github.dreamhead.moco.Moco.template;
@@ -321,9 +322,9 @@ public class MocoTemplateTest extends AbstractMocoHttpTest {
         server.request(by(uri("/template"))).response(template("${now('yyyy-MM-dd')}"));
 
         running(server, () -> {
-            Date date = new Date();
-            SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
-            assertThat(helper.get(remoteUrl("/template")), is(format.format(date)));
+            final ZonedDateTime now = ZonedDateTime.now();
+            final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+            assertThat(helper.get(remoteUrl("/template")), is(formatter.format(now)));
         });
     }
 
@@ -498,6 +499,14 @@ public class MocoTemplateTest extends AbstractMocoHttpTest {
     }
 
     @Test
+    public void should_return_xml_with_declaration() throws Exception {
+        server.request(by(uri("/template"))).response(template("${req.xml.parameter.id}"));
+        running(server, () -> assertThat(helper.postContent(remoteUrl("/template"),
+                "<?xml version=\"1.0\" encoding=\"UTF-8\"?><request><parameter><id>1</id></parameter></request>"),
+                is("1")));
+    }
+
+    @Test
     public void should_return_bad_request_for_unknown_xml() throws Exception {
         server.request(by(uri("/template"))).response(template("${req.xml.parameter.id}"));
         running(server, () -> assertThat(helper.postForResponse(remoteUrl("/template"), "foo").getCode(), is(400)));
@@ -509,6 +518,18 @@ public class MocoTemplateTest extends AbstractMocoHttpTest {
 
         running(server, () -> {
             assertThat(helper.get(remoteUrl("/template")), is("127.0.0.1"));
+        });
+    }
+
+    @Test
+    public void should_generate_response_with_path() throws Exception {
+        server.request(path(uri("/path/{path}"))).response(template("${req.path.path}"));
+        server.request(path(uri("/path/{path}/sub/{sub}"))).response(template("${req.path.path}/${req.path.sub}"));
+
+        running(server, () -> {
+            assertThat(helper.get(remoteUrl("/path/hello")), is("hello"));
+            assertThat(helper.get(remoteUrl("/path/foo")), is("foo"));
+            assertThat(helper.get(remoteUrl("/path/foo/sub/bar")), is("foo/bar"));
         });
     }
 }