+++ moco-doc/ReleaseNotes.md
@@ -1,8 +1,9 @@
-# Release
+# Release 1.4.0 (1-Dec-2022)
 ## APIs
 * `path` is introduced for uri
 * `req.xml` is supported in template
 * `req.client.address` is supported in template
+* `req.path` is supported in template
 
 ## CLI
 * add `-q` for quiet mode in standalone server  
@@ -10,7 +11,6 @@
 ## Implementation Details
 * enhance log with query parameters
 * event action log is supported
-* reduce standalone jar size
 
 # Release 1.3.0 (1-Jan-2022)
 ## APIs
@@ -299,7 +299,7 @@ the first public release of Moco
 
 ## Usage
 * API
-* stanalone server
+* standalone server
 * shell
 * maven plugin (external)
 * gradle plugin (external)